requirements.txt
requests
python-dotenv
Pillow
IPython
unstructured
google-generativeai
langchain
langchain-google-genai
langchain-core
chromadb
PyMuPDF